server/requirements.txt
Django
django-rest-framework
djangorestframework-simplejwt